// src/models/Materials.ts
export interface MaterialsModel {
    title: string
    slug: string
    description: string
    categories: string[]
    supplier: string
    cost: number
    images: string[]
    mass: string
    importDate: string
    expiryDate: string
    createdAt: string
    updatedAt: string
    isDeleted: boolean
    _id: string
    __v: number
}